Removed int64_t and float variants of BC_Pot
Use variants with double and int if possible
Use SPDX-License tags

+++ cinelerra/pluginaclientlad.C
@@ -243,8 +243,8 @@ PluginACLientFLinear::PluginACLientFLinear(PluginAClientLAD *plugin,
 	int x,
 	int y,
 	LADSPA_Data *output,
-	float min,
-	float max)
+	double min,
+	double max)
  : BC_FPot(x, y, *output, min, max)
 {
 	this->plugin = plugin;
